backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
cryptography==41.0.7
bcrypt==4.1.2
Flask-Limiter==3.5.0
psycopg2-binary==2.9.7
